refactor(back): :truck: rename back into rest-server in back dir
