+++ rest-server/Dockerfile
@@ -9,14 +9,14 @@
 
 ##########################
 ### Test dev container ###
-# $ docker image build --target dev -t chess_back_dev ./back
-# $ docker container run -p 8080:8080 chess_back_dev
+# $ docker image build --target dev -t chess_rest_server_dev ./back
+# $ docker container run -p 8080:8080 chess_rest_server_dev
 
 ##########################
 ### Test prod container ##
-# $ docker image build --target prod -t colinfaivre/chess_back_prod ./back
-# $ docker container run -p 8080:8080 chess_back_prod
-# $ docker image push colinfaivre/chess_back_prod
+# $ docker image build --target prod -t colinfaivre/chess_rest_server_prod ./back
+# $ docker container run -p 8080:8080 chess_rest_server_prod
+# $ docker image push colinfaivre/chess_rest_server_prod
 
 FROM golang:1.23-alpine AS base
 WORKDIR /app
