+++ rest-server/controller/authentication.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/octodrome/chess/go-rest-api-poc/helper"
-	"github.com/octodrome/chess/go-rest-api-poc/model"
+	"github.com/octodrome/chess/rest-server/helper"
+	"github.com/octodrome/chess/rest-server/model"
 )
 
 func Register(context *gin.Context) {
