+++ rest-server/controller/user_test.go
@@ -7,8 +7,8 @@ import (
 	"net/url"
 
 	"github.com/gin-gonic/gin"
-	"github.com/octodrome/chess/go-rest-api-poc/controller"
-	"github.com/octodrome/chess/go-rest-api-poc/model"
+	"github.com/octodrome/chess/rest-server/controller"
+	"github.com/octodrome/chess/rest-server/model"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
